% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_run.R
\name{stat_run}
\alias{stat_run}
\title{Stat run}
\usage{
stat_run(mapping = NULL, data = NULL, geom = "path",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, run_length = 1, run_fill_value = NULL,
  run_fill_step = 1, run_fun = base::cumsum, x_run_fill_value = NULL,
  x_run_fun = NULL, y_run_fill_value = NULL, y_run_fun = NULL, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or
\code{\link{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If \code{NULL}, the default, the data is inherited from the plot
   data as specified in the call to \code{\link{ggplot}}.

   A \code{data.frame}, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   \code{\link{fortify}} for which variables will be created.

   A \code{function} will be called with a single argument,
   the plot data. The return value must be a \code{data.frame.}, and
   will be used as the layer data.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behavior from
the default plot specification, e.g. \code{\link{borders}}.}

\item{run_length}{The number of epochs to average over}

\item{run_fill_value}{Fill missing epochs with this value}

\item{run_fill_step}{Fill in the \code{t} aesthetic with this cadance}

\item{run_fun}{The running function. Defaults to \code{base::cumsum}}

\item{x_run_fill_value}{x-specific fill value. Defaults to \code{run_fill_value}}

\item{x_run_fun}{x-specific running function. Defaults to \code{run_fun}}

\item{y_run_fill_value}{y-specific fill value. Defaults to \code{run_fill_value}}

\item{y_run_fun}{y-specific running function. Defaults to \code{run_fun}}
}
\description{
Stat run ab absurdum
}
\examples{
set.seed(101)
wip

}
